Changed how versioning works and wcm publish now checks for overriding
When a component is downloaded it is seperated by a hyphen, name on the left side version on the right. There is no more version parser because it would cause errors if the version field of the yaml had a 'v' in it.

publish will also warn the user if they are about to override a pre existing component